+++ ChangeLog.md
@@ -1,4 +1,12 @@
+* 122 [func] godfryd
+
+    Migrated command line processing in agent from jessevdk/go-flags
+    to urfave/cli/v2. Thanks to this it is possible to define commands
+    in command line. Previously only switches were possible in command
+    line. This is a preparation for new agent command: register.
+
 Stork 0.14.0 released on 2020-12-09.
+    (Gitlab #468)
 
 * 121 [func] marcin